Src/endpoints

+++ src/endpoints.js
@@ -97,6 +97,9 @@ app.get('/users',  requireAuth, async (req, res) => {
 app.get('/users/:id',  requireAuth, async (req, res) => {
     try {
         const { id } = req.params;
+        if (isNaN(id)) {
+            return res.status(400).json({ error: "El ID debe ser un número" });
+        }
         const query = 'SELECT * FROM users WHERE id = $1';
         const result = await pool.query(query, [id]);
 
@@ -165,6 +168,9 @@ app.post('/users', requireAuth, async (req, res) => {
 app.put('/users/:id',  requireAuth, async (req, res) => {
     try {
         const { id } = req.params;
+        if (isNaN(id)) {
+            return res.status(400).json({ error: "El ID debe ser un número" });
+        }
         const { username, email } = req.body;
 
         // Validación 1: Campos vacíos
@@ -234,6 +240,9 @@ app.put('/users/:id',  requireAuth, async (req, res) => {
 app.delete('/users/:id', requireAuth, async (req, res) => {
     try {
         const { id } = req.params;
+        if (isNaN(id)) {
+            return res.status(400).json({ error: "El ID debe ser un número" });
+        }
         const query = 'DELETE FROM users WHERE id = $1';
         const result = await pool.query(query, [id]);
         if (result.rowCount === 0) {
